requirements.txt
fastapi
uvicorn[standard]
ffmpeg-python
pydantic
python-multipart